Moved fetchTasks to AppModalComponent. Moved task detail handler to AppModalComponent.
